fix(daemon): move withEnv helper out of production sources
Sonar still scored the colocated test helper under src/, and finalize could
not re-analyze fork PRs to pick up exclusions. Relocate withEnv under
tests/helpers (already coverage-excluded) and resolve fork PR numbers by
head OID in finalize.
