+++ sonar-project.properties
@@ -13,8 +13,10 @@ sonar.organization=redhat-developer
 sonar.projectKey=redhat-developer_abbenay
 sonar.sources=packages/daemon/src/,packages/vscode/src/
 sonar.tests=packages/daemon/src/,packages/daemon/tests/,packages/vscode/src/test/
-# Keep source/test sets disjoint: co-located *.test.ts (and test helpers like
-# test-env.ts) live under source roots but must be classified only as tests.
-sonar.exclusions=**/node_modules/**,**/dist/**,**/coverage/**,packages/proto-ts/**,packages/python/**,**/*.test.ts,**/test-env.ts
-sonar.test.inclusions=**/*.test.ts,**/test-env.ts
+# Keep source/test sets disjoint: co-located *.test.ts live under source roots
+# but must be classified only as tests (not main sources).
+sonar.exclusions=**/node_modules/**,**/dist/**,**/coverage/**,packages/proto-ts/**,packages/python/**,**/*.test.ts
+sonar.test.inclusions=**/*.test.ts
+# Test helpers under packages/*/tests must never count toward coverage gates.
+sonar.coverage.exclusions=**/*.test.ts,**/tests/helpers/**
 sonar.verbose=false
